factupid/services/api_files/requirements.txt
fastapi
uvicorn
python-multipart
boto3           # luego lo usamos para MinIO
psycopg2-binary # PostgreSQL